requirements.txt
# Requisitos do Sistema de Automação Dark Content

# APIs de IA
openai>=1.0.0

# YouTube Upload (opcional)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Utilitários
requests>=2.28.0
